# Only await cache#get if it's a promise